move availabilty store tests out of overseer package
